pyproject.toml
[tool.poetry]
name = "retail-gravitation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9.17"
streamlit = "^1.26.0"
streamlit-folium = "^0.13.0"
ipykernel = "^6.25.2"
geopandas = "^0.13.2"
matplotlib = "^3.7.3"
mapclassify = "^2.6.0"
pyrosm = "^0.6.1"
shapely = "1.8.5"
prefect = "^2.13.7"
python-dotenv = "^1.0.0"
s3fs = "^2023.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
